# Conflicts:
#	src/main/java/com/incountry/residence/sdk/tools/http/impl/OAuthTokenClient.java

+++ src/integration/java/com/incountry/residence/sdk/StorageIntegrationTest.java
@@ -56,8 +56,8 @@ public class StorageIntegrationTest {
     private static final String PROFILE_KEY = "ProfileKey" + TEMP;
     private static final String KEY_2 = "Key2" + TEMP;
     private static final String KEY_3 = "Key3" + TEMP;
-    private static final Integer BATCH_WRITE_RANGE_KEY = 2;
-    private static final Integer WRITE_RANGE_KEY = 1;
+    private static final Long BATCH_WRITE_RANGE_KEY = 2L;
+    private static final Long WRITE_RANGE_KEY = 1L;
     private static final String RECORD_BODY = "test";
     private static final Integer POOL_SIZE = 5;
